Fix compile error with optional error function parameter on Zig 0.9.1+
`MockRepository.Instance.init` was causing the compiler to panic on zig 0.9.1 and the latest master, owing to it taking an optional error as a parameter. Runtime calls to such functions do compile, but comptime calls (such as the ones that create the global mock repository instances in memory.zig) do not.

I'm still working on a reduced test case for this compiler bug. To unblock myself until it's fixed, I've simplified MockRepository to take a boolean `read_should_fail` parameter instead, which is sufficient for the current unit tests.