Move renderBufferToSurface from Video to Machine
This removes the need for Video to know anything about Hosts and the need for Hosts to know anything about Video; instead, hosts only ever need to talk to the machine instance directly.

This change reduces the surface area of types we need to expose from the library, and better respects the Law of Demeter. It is a prelude to adding more debugging hooks into Machine that notify the host whenever certain machine events occur (such as game part changes, polygon draw calls, or perhaps even register writes).