Rename resource reader struct types
This is the beginning of a codebase-wide reorganization of how types are named and imported.

In the previous convention, a file imported a source file as the base namespace for the type it wanted to access, and then the accessed individual types within it via `Namespace.Instance` or `Namespace.Enum`. This had the problem that Zig would set the canonical name of the type to "Instance" or "Enum" in debug messages.

Under the new convention:
1. all types will be named according to their specific purpose instead of their form;
2. when a source file contains a single obvious "main" type then any other public constants will be exported under that type instead of at the top level;
2. when a source file contains a single obvious "main" type then other files will import that type directly from the source file, e.g. `const MyType = @import("mytype.zig").Mytype`.

This way, will have intrinsic meaningful names that won't collide with the types in other files, and can do away with a base namespace.

(There is a slightly different convention in Zig's own codebase, where files which contain a single struct type will define that struct's fields and methods at the root level of the file itself, without a wrapping `struct {}` statement. This simplifies importing to just `const MyStructType = @import("mystructtype.zig")`. But I don't like this convention because it's specific to structs: it doesn't work for union or enum types, which have to be imported using e.g. `const MyEnumType = @import("myenumtype.zig").MyEnumType`, and I'd rather use that style consistently across all types.)