Move Host to new code organization convention
This is where the import conventions start to break down: MockHost is not a struct, but a function that generates a struct, so we cannot namespace non-generic constants (such as the default host implementation) under it.

Instead we must import the base file and access both the constant and the generator function from within that namespace.