Move Opcode to new code organization convention
This is the first enum that has been migrated to this style. Previously, the raw (as-seen-in-bytecode) type of the enum would live alongside the enum declaration; going forward, the raw type will be exposed as a constant within the enum type. The internal representation of the enum does not need to conform to the raw type so that constraint will be removed.