Convert ResourceID type into non-exhaustive enum
This introduced a similar effort to RegisterID. It increases the friction at sites that create ResourceIDs or consume them as array indices. In return, it prevents arbitrary integers from being erroneously consumed as ResourceIDs (e.g. by mixing up the order of function arguments), and provides a namespace under which constants and helper methods can be grouped.

I am still unsure that the friction of this pattern is worth the benefit.