Make program counter 16-bit
Previously, we parsed 16-bit addresses from bytecode but used a native integer type internally for the program counter, stack and thread locations. This theoretically supported programs of arbitrary size.

However, once we support serializing the game state, this could have led to crashes when loading a save state on a machine with a smaller integer size than the one that recorded it. And we don't need to support arbitrarily large programs, because the original game's bytecode couldn't exceed the capacity of a 16-bit integer.

The virtual machine now uses a 16-bit program counter internally and enforces at runtime that programs cannot exceed that length.