Add bufferChanged callback to notify host of draw calls
This is intended for emulator debug features that may want to display the individual results of each draw call.

Not yet addressed is how to notify the host of palette changes, which effectively modify the content of all 4 buffers at once. One option is to send a notification for each buffer, but that could be suboptimal for the emulator. Another option would be to define a separate callback for this specific operation, or more genericlaly for operations that affect all buffers at once.