WIP implementation of dynamic dispatch table creation
Through some tortuous compile-time metaprogramming, we can mostly eliminate the boilerplate in the "fat pointer" pattern: helpers can generate the necessary type-erased vtable type and wrappers based on a type-aware "template" struct of dispatch functions.

This works beautifully in the test and benchmark targets, but introduces a compile error in the run target because of a circular dependency in the SDL engine. This occurs under both 0.9.1 and 0.10.0-dev.3685+dae7aeb33. I'm not clear if this is a compiler bug, an actual thing I can fix, or a fundamental limitation I can't fix.