Create machine
SDLEngine.runUntilExist now creates a machine at the start of the run loop and destroys it at the end, so the is never stored into the state of the machine itself. This avoids the compile-time circular dependency that was triggered by the dynamic vtable generation.

(It's also tidier conceptually: the machine's state should not be reused between runs.)

In theory this should also fix the bug where we can't stack-allocate the SDLEngine, since the struct is no longer storing pointers to itself into the machine before it gets copied into the final memory location. In practice this still breaks because the game_dir pointer is being copied by the resource directory, so further work is still needed there.