Make vm files load public constants from top level
This pattern will be applied across all domains in the library code. Getting rid of the mix of `const Type = anotherworld.vm.Type` and `const Type = @import("localfile.zig").Type` simplifies reorganizing the codebase in future, as well as helping enforce that the parameter types of public functions are themselves public constants.