Abortive modelling of empty descriptors
The Another World game files include a number of 0-length resources which are tagged as sounds. I figured these were actually unused markers that were squatting the same namespace as sounds; and made the resource loading layer validate that you weren't attempting to load a 0-length resource, instead treating it as if you had specified an ID that didn't exist.

Unfortunately the original game does try to *play* at least one of these 0-length resources as a sound, even though it never loads them; this broke the validation I'd put in to ensure you couldn't try to play a resource of the wrong type as a sound.

In the end, I decided the extra indirection of a separate empty resource type wasn't worth the bad ergonomics and silencing of warnings, and decided to keep let the sound effect parser catch attempts to play a loaded 0-length sound instead.