Rough cut of music player
This player takes a music track and a repository of loaded audio resources (expected to be a Memory instance in the VM, but can be stubbed for testing), and iterates over the patterns of the track in sequence playing channel events. Currently it just logs the events, but eventually it will need to talk to both a mixer and a bank of registers.