Change music delay to tempo
I realized once I began working on the mixer that "delay" was actually the interval between beats of the track, not the delay before the track begins. It also became clear that the set-tempo opcode was only intended to affect the current music track, and was not applied to the next track played if nothing was already playing.