Clamp FrequencyID to the number of preset frequencies
The validation on this type is not satisfying because there's still the possibility to create a value via `@intToEnum()` that's out of bounds; I don't want to define a 40-value enum just to prevent that.

If-and-when ranged integers (https://github.com/ziglang/zig/issues/3806) get implemented, this would be a perfect use case.