Simplify data layout of SoundResource
Previously, sound resources maintained separate slices for the intro and loop sections of the sound. This structure was going to unnecessarily complicate the mixing of the sound: the sound must be played through in its entirety (intro + loop) before looping. It is easier to model and reason about that if we maintain a single slice and reset the mixer's internal counter to the loop point upon reaching the end of the slice.