Preliminary implementation of mixer
This is based heavily on the reference implementation, with some speculation about the hows and whys of its fixed-point mixer. I've noted and speculatively fixed a couple of possible bugs in the original implementation, one of which would have caused looped sounds to loop too early.

There are no tests for this yet, and it's still TBD how to wire it up to the host's audio.