Pump the audio mixer for duration of each frame
This does not yet emit audio to the host system, but goes through the motions of mixing audio data and processing song events.

The reference implementation relied on host timers to handle song events, and SDL audio callbacks to handle mixing. My implementation advances the audio by an equivalent amount each time there's a frame delay, with the idea that the audio state will stay in lockstep with the rest of the VM even if the host system accelerates frame rendering.

However, this ignores the time taken to process bytecode instructions (which would have been taken into account with the Amiga's DMA timers) and I'm not sure what the host system should do with the produced audio if it does accelerate frames; discard it? Modify the frequency itself to imitate a fast-forward effect?