Make Volume into a range-checked nonexhaustive enum
It turns out that Another World's bytecode occasionally defines volumes of 64 which are later clamped down to 63. To avoid parse errors with the original game files, this type must clamp instead of validating that all values fall within range.

Otherwise though, it helped to move this logic off to a type so that volume scaling can be applied in the same way as polygon scaling, abstracting the messy casting away from the business logic.