Reimplement sound sampling with proper looping
This diverges a fair bit from the reference implementation:
- The loop point is correct instead of too early in the sound.
- The sub-byte cursor is taken into account when looping, instead of resetting to the very start of the loop point.
- The channel cursor increases monotonically instead of being rewound: looping is done transparently in the sampler. In theory, this means the cursor will eventually overflow if left to play indefinitely.
- Non-looping samples "fade out" instead of cutting out before the end of the sample.
- Pathological ratios of frequency and sample rate can no longer result in out-of-bounds memory access.
- Samples are not accidentally capped below full volume.

There are still some fixes and optimizations to make: the cursor should be prevented from overflowing, and we should explicitly reset the cursor when looping so that the sampler doesn't have to do as much work on looped samples.

There is also some unexpected round-down behaviour when interpolating off the end of a sample or around the loop point. I'm not sure this is incorrect and it remains to be seen how this affects the sound. But given that the reference implementation seems to have been pretty broken, it should be an improvement in any case.