Return final byte at end of sample data instead of interpolating to 0
I had implemented the interpolation behaviour of unlooped samples with the assumption that silence is represented as 0, and that interpolating to 0 would cause the sample to "trail off".

This is not the case: the values in a stream of PCM audio data mark the absolute position of the speaker at the sampled interval, and speakers produce audio by moving, which means that it's the difference between samples that makes the sound. Silence is represented as *no change* (i.e. movement) in the value over a noticeable period of time.

Thus it may actually be more correct to go back to the reference implementation's approach, which ended playback as soon as the offset reached the last byte.