Use signed 8-bit audio
This better matches the reference implementation, which mixed sound data as signed before converting to unsigned to pass to SDL. I am assuming this means that the original audio files were signed 8-bit data.

This has important implications for volume scaling: thinking of a waveform plotted on a graph, signed values would be scaled toward the centerline of the graph - correctly diminishing the waveform's amplitude - whereas unsigned values would be scaled toward the bottom of the graph, skewing the waveform.