Set marks on music playback
The mark is returned by pointer, from deep inside the music player. This is not a pattern I like or have used elsewhere; I would rather return the mark as a regular return value for the sake of purity and a clearer call convention. But that would force each of the many loops involved in music playback to include boilerplate code to carry the mark value and avoid overwriting it with a no-mark sentinel; the code would be much messier as a result.