Factor rendering-to-surface time into overall frame time
Previously, the time of the last frame was recorded after it had been rendered to the screen. This meant that the time taken to render the frame to the screen was not accounted for in the delay for the next frame, which was arguably correct but caused frames to last longer than the audio generated for them. This contributed to audio buffer starvation, though apparently it was not the only factor; fixing the delay has not cured the problem.